pyproject.toml
[project]
name = "meshinsights-pump-demo"
version = "0.1.0"
description = "MeshInsights Pump Demo"
authors = [
    { name = "Kurt Neuens", email = "kurt.neuens@meshsystems.com" }
]
requires-python = "==3.12.10"

dependencies = [
    "pyodbc==5.2.0",
    "plotly==5.24.1",
    "scipy==1.15.3",
    "scikit-learn==1.7.0",
    "diptest==0.9.0",
    "streamlit==1.49.0",
    "streamlit-authenticator==0.4.2",
    "langchain==0.3.25",
    "langgraph==0.4.2",
    "langchain-mcp-adapters==0.0.10",
    "langchain-openai==0.3.3",
    "langchain-google-genai==2.0.9",
    "langchain-anthropic==0.3.10",
    "langsmith==0.3.13",
    "openai==1.76.0",
    "openai-agents==0.0.13",
    "google-genai==1.8.0",
    "anthropic==0.49.0",
    "pandas==2.3.1",
    "numpy==1.26.4",
    "uvicorn==0.32.0",
    "python-dotenv==1.0.0",
    "pydantic==2.11.3",
    "pydantic-ai==0.1.10",
    "litellm==1.67.5",
    "pyarrow==19.0.0",
    "rapidfuzz==3.13.0",
    "mcp[cli]==1.7.0",
    "resend==2.9.0",
    "fastapi==0.115.12",
    "langserve==0.3.1",
    "matplotlib==3.10.3",
    "azure-common>=1.1.28",
    "azure-core>=1.32.0",
    "azure-identity>=1.13.0",
    "azure-iot-device>=2.12.0",
    "azure-kusto-data>=4.1.4",
    "azure-mgmt-core>=1.4.0",
    "azure-mgmt-resource>=23.0.1",
    "azure-mgmt-resourcehealth>=1.0.0b5",
    "ipykernel==6.29.5",
    "mysql-connector-python==9.4.0",
    "pytest==8.4.1",
    "pandera==0.26.0",
    "tabulate==0.9.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.uv]
package = true
dev-dependencies = [
    "watchdog>=2.1.0",  # Faster Streamlit reloads
    "ipython>=8.0.0",   # Better debugging
]

[tool.pytest.ini_options]
filterwarnings = [
    # Ignore pytz internal use of deprecated utcfromtimestamp on Python 3.12+
    "ignore:datetime\\.datetime\\.utcfromtimestamp\\(\\) is deprecated:DeprecationWarning:pytz(\\..*)?$",
]
